+++ src/components/templates/LoginOnly.tsx
@@ -2,11 +2,8 @@ import React from "react";
 import { navigate } from "gatsby";
 import { useSelector } from "react-redux";
 
-import { useAuthState } from "../../hooks/useAuthState";
-
 export const LoginOnly: React.FC = ({ children }) => {
   const { userDoc, loading } = useSelector((state) => state.auth);
-  useAuthState();
 
   React.useEffect(() => {
     if (!loading && !userDoc) {
