+++ src/pages/setting.tsx
@@ -7,8 +7,9 @@ import { useSelector } from "react-redux";
 import { Layout } from "../components/templates/Layout";
 import { SEO } from "../components/templates/SEO";
 import { LoginOnly } from "../components/templates/LoginOnly";
-import { TwitterConnectButton, TwitterProviderId } from "../components/molecules/TwitterConnectButton";
+import { TwitterConnectButton } from "../components/molecules/TwitterConnectButton";
 import { TwitterUnConnectButton } from "../components/molecules/TwitterUnConnectButton";
+import { TwitterProviderId } from "../hooks/useAuthState";
 import { UserDocType } from "../types/userDoc";
 
 const tweetTimeOptions = [...Array(24).keys()].map((i) => {
